backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.7.0
pinecone-client>=4.1.0
firebase-admin>=6.5.0
pydantic>=2.7.0
python-Levenshtein>=0.25.0
httpx>=0.27.0
python-dotenv>=1.0.0
